regex: the alphabet becomes the code point
Everything the `u` flag changes follows from one substitution — 22.2.1's
alphabet stops being the UTF-16 code unit and becomes the code point — so the
work is finding every place that assumed otherwise. `.` and a class consume a
pair as one character, `search` steps by AdvanceStringIndex rather than ++at
(without which a match could begin between the halves of a pair, and
`/\uDE00/u` would answer true of a whole emoji), and every lastIndex cursor in
the string methods goes through the same step.

Ranges stay code-point ranges in both modes: the two alphabets nest, and a
pattern without `u` can spell no endpoint above 0xFFFF. Only the complement can
tell them apart, because it is the one operation meaning "everything else" — so
that is the only one taking a ceiling, and it is what keeps `\D` under `u` from
excluding the astral plane.

Backward matching had to learn the same lesson the forward direction did, one
commit after gaining a direction at all: it reads the code point before the
position, which means a trail surrogate at pos-1 sends it back two units. The
test that proves it is the one a unit-based read gets wrong rather than the one
it gets right — `/(?<=\uDE00)x/u` is false of an emoji followed by x, where
without `u` it is true.

`\p{...}` and `u` with `i` stay refused by name. The second is new: under both
flags Canonicalize switches to simple case folding, where U+017F folds to `s`
rather than to itself, so the table that just landed would be wrong there in a
way nothing would loudly catch.

+++ src/regex/pattern.h
@@ -20,14 +20,16 @@ enum class NodeKind : uint8_t {
     // An Alternative: the terms, in source order. An empty one matches the
     // empty string, which is what makes `(a|)` legal.
     Sequence,
-    // One code unit, already canonicalized when `i` is set.
+    // One character — a code unit without `u`, a code POINT with it — already
+    // canonicalized when `i` is set.
     Char,
     // A CharacterClass, or a class escape used on its own. `ranges` is
     // normalized; `negated` is `[^...]`, which is NOT the same as complementing
     // the ranges once `i` is in play (22.2.2.7.1 inverts the ANSWER, after the
     // canonicalizing membership test).
     Class,
-    // `.` — every unit but the four line terminators, unless `s` is set.
+    // `.` — every character but the four line terminators, unless `s` is set.
+    // Under `u` that is one code point, so `/^.$/u` matches an astral one.
     Dot,
     // `^`, `$`, `\b`, `\B`.
     Assertion,
@@ -59,7 +61,7 @@ using NodePtr = std::unique_ptr<Node>;
 struct Node {
     NodeKind kind = NodeKind::Sequence;
 
-    uint16_t ch = 0;                  // Char
+    uint32_t ch = 0;                  // Char
     RangeList ranges;                 // Class
     bool negated = false;             // Class
     AssertionKind assertion = AssertionKind::Start;
@@ -77,10 +79,12 @@ struct Node {
     // `/(?:(a)|b)*/.exec("ab")[1]` undefined rather than "a".
     uint32_t firstCapture = 0;
     uint32_t captureCount = 0;
-    // Whether the atom consumes exactly one unit and captures nothing, so the
-    // repetition can be counted and backtracked over ITERATIVELY. Without this
-    // every `.*` would recurse once per input unit and blow the stack on a
-    // string of any size.
+    // Whether the atom consumes exactly one CHARACTER and captures nothing, so
+    // the repetition can be counted and backtracked over ITERATIVELY. Without
+    // this every `.*` would recurse once per input character and blow the stack
+    // on a string of any size. A character is not a fixed number of units under
+    // `u`, so the count is walked back one character at a time rather than
+    // subtracted from the index.
     bool simpleAtom = false;
 
     std::vector<NodePtr> children;
